# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "CTS OCR API",
    "description": "Cheque Truncation System OCR (Optical Character Recognition) - Mobile and Web API Documentation",
    "version": "1.0.0",
    "contact": {
      "name": "CTS Team",
      "email": "support@cts.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5002",
      "description": "Development Server"
    },
    {
      "url": "http://182.76.112.78:91",
      "description": "Production Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Authorization header using the Bearer scheme"
      }
    },
    "schemas": {
...
      "BatchCreationRequest": {
        "type": "object",
        "required": [
          "BranchNationID",
          "ClearingID",
          "ChequeCount",
          "TotalAmount",
          "UserID",
          "PresentingDate",
          "BankID",
          "IFSCCode",
          "presentingMICR",
          "scannerID",
          "gridId"
        ],
        "properties": {
          "BranchNationID": {
            "type": "integer",
            "example": 13
          },
          "ClearingID": {
            "type": "string",
            "example": "01"
          },
          "ChequeCount": {
            "type": "integer",
            "example": 4
          },
          "TotalAmount": {
            "type": "number",
            "format": "double",
            "example": 110.50
          },
          "UserID": {
            "type": "string",
            "example": "demo"
          },
          "PresentingDate": {
            "type": "string",
            "format": "date",
            "example": "2026-05-23"
          },
          "BankID": {
            "type": "integer",
            "example": 3
          },
          "IFSCCode": {
            "type": "string",
            "example": "RATN0000050"
          },
          "presentingMICR": {
            "type": "string",
            "example": "411176003"
          },
          "scannerID": {
            "type": "string",
            "example": "70002"
          },
          "is_DepositeSlip": {
            "type": "boolean",
            "example": false
          },
          "is_SubBankMember": {
            "type": "boolean",
            "example": false
          },
          "gridId": {
            "type": "string",
            "example": "01"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/cts/createBatch": {
      "post": {
        "tags": ["Batch Management"],
        "summary": "Create Batch",
        "description": "Create batch using Oracle database (DEPRECATED - use create_newBatch)",
        "operationId": "batchCreation",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchCreationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Batch created successfully"
          },
          "400": {
            "description": "Invalid batch data"
          }
        },
        "x-status": "DEPRECATED"
      }
    },
    "/cts/create_newBatch": {
      "post": {
        "tags": ["Batch Management"],
        "summary": "Create Batch (MSSQL)",
        "description": "Create batch using Microsoft SQL Server - NEW VERSION",
        "operationId": "batchCreationwithmssql",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchCreationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Batch created successfully in MSSQL"
          },
          "400": {
            "description": "Invalid batch data"
          }
        },
        "x-status": "ACTIVE"
      }
    },
    "/cts/getAllScanDetailsWithBatchid": {
      "post": {
        "tags": ["Batch Management"],
        "summary": "Get Scan Details by Batch ID",
        "description": "Retrieve all scan details for a specific batch",
        "operationId": "getAllScanDetailsWithBatchid",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "batchId": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Scan details retrieved successfully"
          }
        },
        "x-status": "ACTIVE"
      }
    },
...
  }
}